% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FouierCSVToPlot}
\alias{FouierCSVToPlot}
\title{Draw a plot from CSV file}
\usage{
FouierCSVToPlot(data.path = NULL, dir = NULL, freq = c(400, 5000),
  title = "", names = NULL, sqrt.trans = TRUE)
}
\arguments{
\item{data.path}{path to file}

\item{dir}{directory with files}

\item{freq}{frequencies to show on x axis}

\item{title}{title of the plot}

\item{names}{names for plot legend}

\item{sqrt.trans}{square root axist transformation}
}
\value{
An object of class ggplot.
}
\description{
Function \code{FouierCSVToPlot} draws a plot.
}
