% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierAggregatePlot}
\alias{FourierAggregatePlot}
\title{Saving plots from Mean Data}
\usage{
FourierAggregatePlot(data, freq = c(400, 5000), sqrt.trans = TRUE)
}
\arguments{
\item{data}{data generated by \code{FourierMeans} function}

\item{freq}{frequencies on x axis}

\item{sqrt.trans}{square root axis}
}
\value{
Saved plot
}
\description{
Function \code{FouierAggrDir} saves plots from data generated by \code{FourierMeans} function.
}
