% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierDraw}
\alias{FourierDraw}
\title{Draw a plot}
\usage{
FourierDraw(data.means, freq = c(400, 4100), title = "")
}
\arguments{
\item{data.means}{data.frame with means for frequencies}

\item{freq}{frequencies to show on x axis}

\item{title}{title of the plot}
}
\value{
An object of class ggplot.
}
\description{
Function \code{FouierPlotMany} draws a plot.
}
