% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierDrawPlotly}
\alias{FourierDrawPlotly}
\title{Draw a FourierPlot using plotly}
\usage{
FourierDrawPlotly(data.means, freq = c(400, 5000), title = "",
  log.trans = TRUE)
}
\arguments{
\item{data.means}{data.frame with means for frequencies}

\item{freq}{frequencies to show on x axis}

\item{title}{title of the plot}

\item{log.trans}{log axist transformation}
}
\value{
An object of class ggplot.
}
\description{
Function \code{FouierPlotPlotly} draws a plot.
}
