% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierMeans}
\alias{FourierMeans}
\title{Means for frequencies}
\usage{
FourierMeans(data, names = NULL)
}
\arguments{
\item{data}{data.table or a list of data.tables to compute means for frequencies}

\item{names}{The names corresponding to the samples}
}
\value{
An object of class data.frame containing means for specified frequencies.
}
\description{
Function \code{FouierMeans} computes means for specified frequencies.
}
