% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierReadCSV}
\alias{FourierReadCSV}
\title{Reading csv file}
\usage{
FourierReadCSV(data.paths = NULL, dir)
}
\arguments{
\item{data.paths}{path to file or vector of paths to files}

\item{dir}{directory with files}
}
\value{
An object of class data.frame, containing values of ... for specified frequencies.
}
\description{
Function \code{FourierReadCSV} reads data from CSV file with format... see....
}
