% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{GenerateReport}
\alias{GenerateReport}
\title{Generating reports form files}
\usage{
GenerateReport(data.path, freq = c(400, 4100), title = NA,
  plot_all = TRUE, plot_single = FALSE)
}
\arguments{
\item{data.path}{vector of paths to files.}

\item{freq}{frequencies to show on x axis}

\item{title}{a vecor of titles of the plots}

\item{plot_all}{show plot with all samples}

\item{plot_single}{show single plots}
}
\value{
A pdf report.
}
\description{
Function \code{GenerateReport} generates report from CSV files with format... see....
}
