% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyRun.R
\name{ShinyRun}
\alias{ShinyRun}
\title{Running Shiny appllication}
\usage{
ShinyRun()
}
\value{
Run a Shiny app.
}
\description{
Function \code{FourierReadCSV} reads data from CSV file with format... see....
}
