% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fourier.R
\name{FourierReadCSV}
\alias{FourierReadCSV}
\title{Reading csv file}
\usage{
FourierReadCSV(data.path)
}
\arguments{
\item{data.path}{path to file}
}
\value{
An object of class data.frame, containing values of ... for specified frequencies.
}
\description{
Function \code{FourierReadCSV} reads data from CSV file with format... see....
}
